Complete Search Book Screen

+++ app/src/main/java/com/example/myreader/model/Book.kt
@@ -3,5 +3,5 @@ package com.example.myreader.model
 data class Book(
     val items: List<Item>,
     val kind: String,
-    val totalItems: Int
+    val totalItems: Long
 )
